Move init logic into own module and use internal logger
