+++ lib/init/index.js
@@ -1,10 +1,4 @@
-import { existsSync } from 'fs-extra';
 import chalk from 'chalk';
-import cleanupFiles from './utils/clean-up-files';
-import { copyNetwork, copyWithFramework } from './utils/copy-with';
-
-const CWD = process.cwd();
-const BIGTEST_DIR = `${CWD}/bigtest`;
 
 function builder(yargs) {
   yargs
@@ -25,39 +19,7 @@ function builder(yargs) {
 }
 
 async function handler(argv) {
-  let { network, appFramework } = argv;
-  let bigtestDirExists = existsSync(BIGTEST_DIR);
-  let networkDirExists = existsSync(`${BIGTEST_DIR}/network`);
-  let isCreatingNetwork = !networkDirExists && network;
-
-  if (bigtestDirExists && !isCreatingNetwork) {
-    console.info(`\nLooks like BigTest is already initialized\n`);
-
-    return;
-  }
-
-  if (bigtestDirExists && isCreatingNetwork) {
-    await copyNetwork(CWD, appFramework);
-
-    console.info('\n@bigtest/network has been initialized\n');
-
-    return;
-  }
-
-  try {
-    await copyWithFramework(CWD, appFramework, network).then(
-      ({ needsNetwork }) => {
-        let networkMessage = needsNetwork ? 'and @bigtest/mirage' : '';
-
-        console.info(
-          `\nBigTest has been initialized with @bigtest/${appFramework} ${networkMessage}\n`
-        );
-      }
-    );
-  } catch (error) {
-    console.log(`Initialize failed :( \n${error}`);
-    await cleanupFiles('bigtest', CWD);
-  }
+  await require('./init').default(argv);
 }
 
 module.exports = {
